==> depreciated.Rmd/10.climate.niche.Rmd <==
---
title: "10.climate.niche"
author: "Daniele Filiault"
date: "7/1/2019"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## 1. Rationale.

So it looks like N/S selection pressure is significant, but not 100% of differences in local adaptation
Can we identify others with climate parameters?

```{r data prep}
#accession data
c.acc <- read.table('./data/worldclim_swedish_acc.txt',stringsAsFactors=FALSE, header=TRUE, sep="\t")  ### 6 accessions don't have data.
#experiment data
c.exp <- read.table('./data/worldclim_swedish_exp.txt', stringsAsFactors=FALSE, header=TRUE, sep="\t")

### need to merge these so can use together
### columns: name, region, type, lat, long, climate.params....

### accession data first
c.acc <- c.acc[, -2]
colnames(c.acc)[1] <- "name"
colnames(c.acc)[3] <- "type"
c.acc[,3] <- "acc"
## remove 6 accessions with NA.  Can I get these data somewhere???
c.acc <- c.acc[is.na(c.acc$bio_1)==FALSE,]
rownames(c.acc) <- c.acc$name

### experimental data
c.exp <- c.exp[c.exp[,1]=="Common garden",]
c.exp <- c.exp[,-1]
c.exp$region <- c("N Sweden", "N Sweden","S Sweden", "S Sweden")
c.exp$type <- "exp"
c.exp <- c.exp[,c(1,72,73,2:71)]
colnames(c.exp)[1] <- "name"
c.exp <- c.exp[,c(1:3,5,4,6:73)]
colnames(c.exp)[4] <- "long"
rownames(c.exp) <- c.exp$name

### OK, these look safe to rbind.

c.dat <- rbind(c.exp, c.acc)
```

## 2. PCA analysis of accession data

can we define the localness of accessions by "climate space"?

```{r acc climate space PCA}
library(corrplot)
corrplot(cor(c.dat[,4:73]), method="color", type="upper", diag=FALSE)

### just look at core 19 bioclim vars
pdf("./figures/PCA.bioclim.vars.pdf", width=10, height=10)
corrplot(cor(c.dat[,4:24]), method="ellipse", type="upper", diag=FALSE)
dev.off()


## clearly a LOT of correlation here.  Good data for PCA

up.dat <- c.dat[,4:73]
acc.pca <- prcomp(up.dat, center=TRUE)
acc.pca
summary(acc.pca)

library(ggfortify)

autoplot(prcomp(up.dat, center=TRUE), data = c.dat, loadings = TRUE, colour="region",size="type")

### only use 19 climate variables

up.dat <- c.dat[,6:24]
rownames(up.dat) <- rownames(c.dat)
mini.pca <- prcomp(up.dat, center=TRUE)
mini.pca
summary(mini.pca)

library(ggfortify)

autoplot(prcomp(up.dat, center=TRUE), data = c.dat, loadings = TRUE, colour="region",size="type")

### I like this analysis using only the 19 climate variables
### let's look if locally adaptive

pcs <- as.data.frame(mini.pca$x)
pcs$id <- rownames(pcs)
```

```{R mini PCA plot by hand for annotation}
pdf("./figures/PCA.accessions.19climvars.pdf", width=8, height=6)
plot(pcs$PC1, pcs$PC2, xlab="PC1 (99.6%)", ylab="PC2 (0.34%)", type="n")
n.accs <- pcs[rownames(pcs)%in%rownames(c.dat[c.dat$region=="N Sweden",]),]
s.accs <- pcs[rownames(pcs)%in%rownames(c.dat[c.dat$region=="S Sweden",]),]
c.accs <- pcs[rownames(pcs)%in%rownames(c.dat[c.dat$region=="C Sweden",]),]
points(n.accs$PC1, n.accs$PC2, pch=19, col="dodgerblue")
points(c.accs$PC1, c.accs$PC2, pch=19, col="purple")
points(s.accs$PC1, s.accs$PC2, pch=19, col="green")
n.exps <- n.accs[rownames(n.accs)%in%rownames(c.dat[c.dat$type=="exp",]),]
s.exps <- s.accs[rownames(s.accs)%in%rownames(c.dat[c.dat$type=="exp",]),]
points(n.exps$PC1, n.exps$PC2, pch=23, cex=2.5, lwd=2)
points(s.exps$PC1, s.exps$PC2, pch=23, cex=2.5, lwd=2)
legend("bottomleft", col=c("green", "purple","blue","black"),pch=c(19,19,19,23), legend=c("Southern lines", "Central lines", "Northern lines", "Experiment sites"))
exp.sites <- rbind(n.exps,s.exps)
text(exp.sites[,1], exp.sites[,2], labels=c("N2","N1","S1","S2"), pos=c(1,3,1,3), offset=1)
dev.off()

### technically, I have also plotted the experiment locations as accessions.  Fix this if using this plot for something real.


```
From the loadings of these PCS:
PC1 -0.9971 for bio_4 (Temperature Seasonality (standard deviation *100))
PC2 -0.8834 for bio_12 (Annual Precipitation)
PC3 0.902 for bio_8 (Mean Temperature of Wettest Quarter)



### relationship between PCs and fitness
Now that we have these various PCs, is there any evidence that being closer to the experiment site in climate space increases fitness of an accession?
```{r climate adapt}
## need fitness data

fitness <- read.table("./data/fitness.fitted.values.txt")
site.pcs <- pcs[1:4,]
rownames(site.pcs) <- c("ADA","RAM","ULL", "RAT")
fitness <- merge(fitness,pcs)
acc.pcs <- fitness[,grep("PC", colnames(fitness))]

## do this by experimental site


all.lm <- as.list(1:nrow(site.pcs))

for(up.s in 1:nrow(site.pcs)){
  up.site <- rownames(site.pcs)[up.s]
  up.s.pcs <- site.pcs[up.s,1:19]
  up.a.pcs <- apply(acc.pcs, 1, function(x){up.s.pcs-x}) # gets pca relative to experiment site
  up.a.pcs <- do.call(rbind, up.a.pcs)
  up.fit <- fitness[,grep(up.site, colnames(fitness))]
  up.fit <- up.fit[,-grep("both", colnames(up.fit))]
  
  ### do lm for first 3 PCs for 2011
  lm2011 <- as.list(1:3)
  for(up.p in 1:3){
    up.pcs <- abs(up.a.pcs[,up.p])  ### take abs value to get relative distance
    up.lm <- lm(up.fit[,1] ~ up.pcs)
    lm2011[[up.p]] <- up.lm
  }
  
  ### do lm for first 3 PCs for 2012
  lm2012 <- as.list(1:3)
  for(up.p in 1:3){
    up.pcs <- abs(up.a.pcs[,up.p])  ### take abs value to get relative distance
    up.lm <- lm(up.fit[,2] ~ up.pcs)
    lm2012[[up.p]] <- up.lm
  }
  
  ### output these
  lmboth <-c(lm2011, lm2012)
  all.lm[[up.s]] <- lmboth
  
}

### pull out pvals, adjr2, and slopes

pvals <- lapply(all.lm, function(x) {lapply(x, function(y){anova(y)[[5]][1]} )})
pvals <- unlist(pvals)

adjR <- lapply(all.lm, function(x) {lapply(x, function(y){summary(y)$adj.r.squared} )})
adjR <- unlist(adjR)

## we want slope to be negative - further away, lower fitness
slope <- lapply(all.lm, function(x) {lapply(x, function(y){y$coefficients[2]} )})
slope <- unlist(slope)

model.sums <- cbind(pvals, adjR, slope)

## these need some rownames before I forget

sites <- as.vector(sapply(rownames(site.pcs),function(x){rep(x,6)}))
pcnum <- rep(c(1:3),8)
year <- rep(c(rep("2011",3), rep("2012",3)),4)
rnames <- paste(sites, pcnum, year, sep="_")
rownames(model.sums) <- rnames

model.sums <- model.sums[order(model.sums[,1]),]

print(model.sums)
```


BIO1 = Annual Mean Temperature
BIO2 = Mean Diurnal Range (Mean of monthly (max temp - min temp))
BIO3 = Isothermality (BIO2/BIO7) (* 100)
BIO4 = Temperature Seasonality (standard deviation *100)
BIO5 = Max Temperature of Warmest Month
BIO6 = Min Temperature of Coldest Month
BIO7 = Temperature Annual Range (BIO5-BIO6)
BIO8 = Mean Temperature of Wettest Quarter
BIO9 = Mean Temperature of Driest Quarter
BIO10 = Mean Temperature of Warmest Quarter
BIO11 = Mean Temperature of Coldest Quarter
BIO12 = Annual Precipitation
BIO13 = Precipitation of Wettest Month
BIO14 = Precipitation of Driest Month
BIO15 = Precipitation Seasonality (Coefficient of Variation)
BIO16 = Precipitation of Wettest Quarter
BIO17 = Precipitation of Driest Quarter
BIO18 = Precipitation of Warmest Quarter
BIO19 = Precipitation of Coldest Quarter
