==> depreciated.Rmd/16.fitness.by.admixture.Rmd <==
---
title: "16.fitness.by.admixture"
author: "Daniele Filiault"
date: "9/16/2019"
output:
  html_document: default
  pdf_document: default
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library("RColorBrewer")
library("ggplot2")

library(ggmap)
library(maps)
library(mapdata)
library("sf")
library("rnaturalearth")
library("rnaturalearthdata")
library("rgeos")
library(viridis)
library(reshape)
library(gridExtra)

group.colors <- brewer.pal(9, "Paired")
ss.col <- group.colors[3]
group.colors[3] <- group.colors[7]
group.colors[7] <- ss.col
ag.colors <-group.colors[c(3,7,4)]

```

## Introduction

Interested in the Central Swedish accessions, who can be seen as "admixed" samples between N and S admixture groups.  Is this true?  Does the proportion of admixture relate to accession fitness in a predictable way in the common gardens?


```{r prep admixture data}
### admixture groups for 1001g data - from the paper:
#We defined nine groups based on these clusters and assigned each individual to a group if more than 60% of its genome
#derived from the corresponding cluster. The 135 individuals not
#matching this criterion were labeled ‘‘Admixed.’’
admix <- read.csv(file="./data/1001genomes-accessionsand1001genomes-admixture.csv", stringsAsFactors=FALSE)
admix.m <- admix[,grep("K9", colnames(admix))]
rownames(admix.m) <- admix$id
#admix.m$id <- rownames(admix.m)

max.a <- apply(admix.m,1, function(x)(max(x, na.rm=TRUE)))
a.group <- rep(NA, nrow(admix.m))
for(up in 1:nrow(admix.m)){
  up.dat <- admix.m[up,]
  a.group[up] <- which(up.dat==max.a[up])-1
}
names(a.group)<- rownames(admix.m)
a.group <- cbind(admix[,c(1:4)], max.a,a.group, admix.m)


## subset for experiment accessions.  Some will be missing.  Should rerun admixture with expanded data?
fit.dat <- read.table("./data/fitness.fitted.values.txt", stringsAsFactors=FALSE)
fit.dat <- fit.dat[,c(1,8:11,14:17)]

a.fit <- merge(a.group,fit.dat)  ### only 186 out of 200 accessions have admixture info

## get info about N/C/S accession
acc.list <- read.table("./data/acc_list.txt", sep="\t", header=TRUE)
colnames(acc.list)[1] <- "id"
a.fit <- merge(a.fit, acc.list[,c(1,3)])
a.fit$region <- droplevels(a.fit$region)

```

## Let's ask some basic questions about admixture groups here

```{R admix group summary}
with(a.fit, table(a.group, region))

## where are these 5 groups?

with(a.group, table(a.group, country))

a.fit <- a.fit[order(a.fit$region,a.fit$a.group, a.fit$max.a),]
hist(a.fit$max.a)
a.fit$region <- factor(a.fit$region, levels=c("S Sweden", "C Sweden", "N Sweden"))
a.fit <- a.fit[order(a.fit$region, a.fit$a.group, a.fit$max.a),]

c.lim <- (which(a.fit$region=="C Sweden"))
s.lim <- (which(a.fit$region=="S Sweden"))
n.lim <- (which(a.fit$region=="N Sweden"))

group.names <- c("W.Europe","Relicts","Germany","N.Sweden","Iberia","Asia","S.Sweden","Italy/E.Europe","C.Europe")
group.colors <- brewer.pal(9, "Paired")
ss.col <- group.colors[3]
group.colors[3] <- group.colors[7]
group.colors[7] <- ss.col

pdf("./figures/admixture.groups.experimental.lines.pdf", width=14, height=7)
par(xpd=TRUE)
barplot(t(as.matrix(a.fit[,7:15])), col=group.colors, main="All experimental lines", legend.text=group.names, xaxt="n",space=0, args.legend=list(title="admixture group",bg="white"), ylab="genome proportion", xlab="collection location in Sweden")
rect(min(c.lim)-1,-0.04,max(c.lim),-0.01, col="purple")
rect(min(s.lim)-1,-0.04,max(s.lim),-0.01, col="green")
rect(min(n.lim)-1,-0.04,max(n.lim),-0.01, col="blue")
mtext("Central",1,line=1,at=mean(c.lim)-0.5)
mtext("Southern",1,line=1,at=mean(s.lim)-0.5)
mtext("Northern",1,line=1,at=mean(n.lim)-0.5)
dev.off()


### boxplots of main admixture groups by accession geography
par(mfcol=c(2,2))
with(a.fit, boxplot(K9.ChrAll.3~region, ylab="prop N.Sweden admixture group"))
with(a.fit, boxplot(K9.ChrAll.6~region, ylab="prop S.Sweden admixture group"))
with(a.fit, boxplot(K9.ChrAll.2~region, ylab="prop German admixture group"))
with(a.fit, boxplot((1-(K9.ChrAll.3+K9.ChrAll.6+K9.ChrAll.2))~region,ylab="prop other admixture groups"))

```

```{R plot geo locations admix groups}
### geographic locations of admixture groups
bc.dat <- read.table("./data/bioclim.v2.200.experimental.lines.txt",stringsAsFactors=FALSE)
colnames(bc.dat)[1] <- "id"
bc.dat <- merge(a.fit, bc.dat[,1:3])

theme_set(theme_bw())
world <- ne_countries(scale = "medium", returnclass = "sf")

sweden <- map('worldHires','Sweden')
sweden.map <- ggplot(data = world) +geom_sf() + coord_sf(xlim = c(sweden$range[1], sweden$range[2]), ylim = c(sweden$range[3], sweden$range[4]), expand = FALSE) + scale_x_continuous(breaks=c(12,16,20,24))

NSweden.map <- sweden.map + geom_point(data=bc.dat, mapping = aes(x = lon, y = lat, colour = K9.ChrAll.3)) + labs(colour = "proportion \nN.Swedish") + scale_color_viridis(option = "C")
SSweden.map <- sweden.map + geom_point(data=bc.dat, mapping = aes(x = lon, y = lat, colour = K9.ChrAll.6)) + labs(colour = "proportion \nS.Swedish") + scale_color_viridis(option = "C")
Ger.map <- sweden.map + geom_point(data=bc.dat, mapping = aes(x = lon, y = lat, colour = K9.ChrAll.2)) + labs(colour = "proportion \nGerman    ") + scale_color_viridis(option = "C")
Other.map <- sweden.map + geom_point(data=bc.dat, mapping = aes(x = lon, y = lat, colour = (1-(K9.ChrAll.3+K9.ChrAll.6+K9.ChrAll.2)))) + labs(colour = "prop Other") + scale_color_viridis(option = "C")
print(NSweden.map)
print(SSweden.map)
print(Ger.map)
print(Other.map)

pdf("./figures/admix.group.Sweden.map.pdf", width=12, height=4.5)
grid.arrange(Ger.map, SSweden.map, NSweden.map, nrow = 1)
dev.off()

```
```{R plot geo locations 1001g three admix groups}
tog.loc <- read.csv("./data/accs_2029_ids_lat_long.modified_MN_SH.csv",stringsAsFactors=FALSE)
colnames(tog.loc)[2] <- "id"

europe.map <- ggplot(data = world) +geom_sf() + coord_sf(xlim = c(-7, 20), ylim = c(47, 65), expand = FALSE) +   scale_x_continuous(breaks=c(0,10,20))

e.dat <- a.group[a.group$a.group%in%c(2,3,6),]
e.dat$admix <- e.dat$max.a<0.6
e.dat$a.group <- factor(e.dat$a.group, levels=c(2,6,3))
e.dat <- merge(e.dat, tog.loc[,c(2,5,6)])
e.dat <- e.dat[e.dat$admix==FALSE,]

e.map <- europe.map + 
  geom_point(dat=e.dat,mapping = aes(x = longitude, y = latitude, color = a.group)) +
  theme_bw() +
  scale_color_manual(values=ag.colors, name="admixture \ngroup", labels=c("German", "S.Swedish", "N.Swedish"))


pdf(file="./figures/3admixturegroup.europe.map.pdf", width=5, height=6)
print(e.map)
dev.off()


```




```{R admix group vs bioclim}
pca.dat <- read.table("./data/experimental.pcas.txt", stringsAsFactors=FALSE)
pcs<- merge(pca.dat,a.fit)
site.pcs <- pca.dat[1:8,]
site.labels <- c("N2-C","N2-S","N3","N1","S2","S2","S3","S1")

NS.pca <- ggplot(pcs, aes(x=PC1, y=PC2, colour=K9.ChrAll.3)) + geom_point() + labs(colour = "proportion \nN.Swedish") + scale_color_viridis(option = "C") + geom_point(data = site.pcs, col = 'black', shape=5, size=3) + geom_text(data=site.pcs, label=site.labels, colour="black", nudge_x=c(51,41,34,34,34,34,34,34), nudge_y=c(0,-8,0,0,0,0,-5,-10))
SS.pca <- ggplot(pcs, aes(x=PC1, y=PC2, colour=K9.ChrAll.6)) + geom_point() + labs(colour = "proportion \nS.Swedish") + scale_color_viridis(option = "C")+ geom_point(data = site.pcs, col = 'black', shape=5, size=3)+ geom_text(data=site.pcs, label=site.labels, colour="black", nudge_x=c(51,41,34,34,34,34,34,34), nudge_y=c(0,-8,0,0,0,0,-5,-10))
GER.pca <- ggplot(pcs, aes(x=PC1, y=PC2, colour=K9.ChrAll.2)) + geom_point() + labs(colour = "proportion \nGerman   ") + scale_color_viridis(option = "C")+ geom_point(data = site.pcs, col = 'black', shape=5, size=3)+ geom_text(data=site.pcs, label=site.labels, colour="black", nudge_x=c(51,41,34,34,34,34,34,34), nudge_y=c(0,-8,0,0,0,0,-5,-10))
OTHER.pca <- ggplot(pcs, aes(x=PC1, y=PC2, colour=(1-(K9.ChrAll.3+K9.ChrAll.6+K9.ChrAll.2)))) + geom_point() + labs(colour = "prop Other") + scale_color_viridis(option = "C")+ geom_point(data = site.pcs, col = 'black', shape=5, size=3)

print(NS.pca)
print(SS.pca)
print(GER.pca)
print(OTHER.pca)

pdf("./figures/admix.group.bioclim.PC.pdf", width=14, height=4.5)
grid.arrange(GER.pca, SS.pca, NS.pca, nrow = 1)
dev.off()
```

So, five admixture groups total here (0,2,3,6,8).  N Sweden seems to be group 3, South predominately group 6.  Both these admixutre groups are predominantly Swedish.  Group 0 is spain/uk, group 2 is germany/usa, group 8 is austria/czech/italian/german.

But many that one would consider "admixed" - kind of 2 sets of admixed in Central lines
1. S.Sweden with Asian
2. N.Sweden with German

Is there a relationship between the proportion of admixture group and fitness at a site (that makes sense with local adaptation)?


### Effect of Proportion of admixture group
```{R fitness by admixture proportion}
fit.columns <- colnames(a.fit)[16:23]

plot.fit.prop <- function(fit.col){
  par(mfcol=c(2,2))
  with(a.fit, plot(K9.ChrAll.3, get(fit.col), type="n", xlab="proportion genome N.Swedish", ylab=paste("fitness",fit.col, sep=" ")))
  with(a.fit[a.fit$region=="N Sweden",], points(K9.ChrAll.3, get(fit.col), col="cadetblue", pch=19))
  with(a.fit[a.fit$region=="S Sweden",], points(K9.ChrAll.3, get(fit.col), col="green", pch=19))
  with(a.fit[a.fit$region=="C Sweden",], points(K9.ChrAll.3, get(fit.col), col="purple", pch=19))
  abline(lm(get(fit.col)~K9.ChrAll.3, data=a.fit))
  abline(lm(get(fit.col)~K9.ChrAll.3, data=a.fit[a.fit$region=="C Sweden",]), col="purple", lty=4)

  with(a.fit, plot(K9.ChrAll.6, get(fit.col), type="n", xlab="proportion of genome S.Swedish", ylab=paste("fitness",fit.col, sep=" ")))
  with(a.fit[a.fit$region=="S Sweden",], points(K9.ChrAll.6, get(fit.col), col="green", pch=19))
  with(a.fit[a.fit$region=="N Sweden",], points(K9.ChrAll.6, get(fit.col), col="cadetblue", pch=19))
  with(a.fit[a.fit$region=="C Sweden",], points(K9.ChrAll.6, get(fit.col), col="purple", pch=19))
  abline(lm(get(fit.col)~K9.ChrAll.6, data=a.fit))
  abline(lm(get(fit.col)~K9.ChrAll.6, data=a.fit[a.fit$region=="C Sweden",]), col="purple", lty=4)

  with(a.fit, plot(K9.ChrAll.2, get(fit.col),type="n", xlab="proportion of genome German", ylab=paste("fitness",fit.col, sep=" ")))
  with(a.fit[a.fit$region=="S Sweden",], points(K9.ChrAll.2, get(fit.col), col="green", pch=19))
  with(a.fit[a.fit$region=="N Sweden",], points(K9.ChrAll.2, get(fit.col), col="cadetblue", pch=19))
  with(a.fit[a.fit$region=="C Sweden",], points(K9.ChrAll.2, get(fit.col), col="purple", pch=19))
  abline(lm(get(fit.col)~K9.ChrAll.2, data=a.fit))
  abline(lm(get(fit.col)~K9.ChrAll.2, data=a.fit[a.fit$region=="C Sweden",]), col="purple", lty=4)

  with(a.fit, plot(1-(K9.ChrAll.3+K9.ChrAll.6), get(fit.col), type="n", xlab="proportion of genome other groups", ylab=paste("fitness",fit.col, sep=" ")))
  with(a.fit[a.fit$region=="N Sweden",], points(1-(K9.ChrAll.3+K9.ChrAll.6+K9.ChrAll.2), get(fit.col), col="cadetblue", pch=19))
  with(a.fit[a.fit$region=="S Sweden",], points(1-(K9.ChrAll.3+K9.ChrAll.6+K9.ChrAll.2), get(fit.col), col="green", pch=19))
  with(a.fit[a.fit$region=="C Sweden",], points(1-(K9.ChrAll.3+K9.ChrAll.6+K9.ChrAll.2), get(fit.col), col="purple", pch=19))
  p.nsw <- 1-(a.fit$K9.ChrAll.3+a.fit$K9.ChrAll.6+a.fit$K9.ChrAll.2)
  abline(lm(get(fit.col)~p.nsw, data=a.fit))
  p.nsw.c <- 1-(a.fit[a.fit$region=="C Sweden",]$K9.ChrAll.3+a.fit[a.fit$region=="C Sweden",]$K9.ChrAll.6+a.fit[a.fit$region=="C Sweden",]$K9.ChrAll.2)
  abline(lm(get(fit.col)~p.nsw.c, data=a.fit[a.fit$region=="C Sweden",]), col="purple", lty=4)
  legend(x="topright", legend=c("North","Central", "South"), col=c("cadetblue","purple","green"), pch=19)
}

for(up.c in fit.columns){
  pdf(file=paste("./figures/fit.by.admix.prop.", up.c, ".pdf", sep=""), width=8, height=6)
  par(mar=c(5,4,1,1))
  plot.fit.prop(fit.col=up.c)
  dev.off()
}

```

What's most important in determining fitness, prop N, prop S, or prop non-Swedish?

These proportions are weird variables (high correlation), and there is probably a better way to do this, but let's just start with plain old linear models like in the above figures

```{R lm fit by prop}

prop.lm <- function(fit.col){
  out.dat <- as.list(1:4)
  out.dat[[1]] <- lm(get(fit.col)~K9.ChrAll.3, data=a.fit)
  out.dat[[2]] <- lm(get(fit.col)~K9.ChrAll.6, data=a.fit)
  out.dat[[3]] <- lm(get(fit.col)~K9.ChrAll.2, data=a.fit)
  p.nsw <- 1-(a.fit$K9.ChrAll.3+a.fit$K9.ChrAll.6+a.fit$K9.ChrAll.2)
  out.dat[[4]] <- lm(get(fit.col)~p.nsw, data=a.fit)
  return(out.dat)
}

### 2011 first
fit.columns <- colnames(a.fit)[grep("2011",colnames(a.fit))]
ada1 <- prop.lm(fit.col=fit.columns[1])
ram1 <- prop.lm(fit.col=fit.columns[2])
ull1 <- prop.lm(fit.col=fit.columns[3])
rat1 <- prop.lm(fit.col=fit.columns[4])

### 2012 second
fit.columns <- colnames(a.fit)[grep("2012",colnames(a.fit))]
ada2 <- prop.lm(fit.col=fit.columns[1])
ram2 <- prop.lm(fit.col=fit.columns[2])
ull2 <- prop.lm(fit.col=fit.columns[3])
rat2 <- prop.lm(fit.col=fit.columns[4])

###### slopes ######
### 2011
ada1.s <- unlist(lapply(ada1, function(x){summary(x)$coefficients[2,1]}))
ram1.s <- unlist(lapply(ram1, function(x){summary(x)$coefficients[2,1]}))
ull1.s <- unlist(lapply(ull1, function(x){summary(x)$coefficients[2,1]}))
rat1.s <- unlist(lapply(rat1, function(x){summary(x)$coefficients[2,1]}))
slope1 <- rbind(ada1.s, ram1.s, ull1.s, rat1.s)
colnames(slope1) <- c("n.swedish", "s.swedish", "german", "other")

### 2012
ada2.s <- unlist(lapply(ada2, function(x){summary(x)$coefficients[2,1]}))
ram2.s <- unlist(lapply(ram2, function(x){summary(x)$coefficients[2,1]}))
ull2.s <- unlist(lapply(ull2, function(x){summary(x)$coefficients[2,1]}))
rat2.s <- unlist(lapply(rat2, function(x){summary(x)$coefficients[2,1]}))
slope2 <- rbind(ada2.s, ram2.s, ull2.s, rat2.s)
colnames(slope2) <- c("n.swedish", "s.swedish", "german", "other")

######### pvals #############
### 2011
ada1.p <- unlist(lapply(ada1, function(x){summary(x)$coefficients[2,4]}))
ram1.p <- unlist(lapply(ram1, function(x){summary(x)$coefficients[2,4]}))
ull1.p <- unlist(lapply(ull1, function(x){summary(x)$coefficients[2,4]}))
rat1.p <- unlist(lapply(rat1, function(x){summary(x)$coefficients[2,4]}))
pval1 <- rbind(ada1.p, ram1.p, ull1.p, rat1.p)
colnames(pval1) <- c("n.swedish", "s.swedish", "german", "other")

### 2012
ada2.p <- unlist(lapply(ada2, function(x){summary(x)$coefficients[2,4]}))
ram2.p <- unlist(lapply(ram2, function(x){summary(x)$coefficients[2,4]}))
ull2.p <- unlist(lapply(ull2, function(x){summary(x)$coefficients[2,4]}))
rat2.p <- unlist(lapply(rat2, function(x){summary(x)$coefficients[2,4]}))
pval2 <- rbind(ada2.p, ram2.p, ull2.p, rat2.p)
colnames(pval2) <- c("n.swedish", "s.swedish", "german", "other")



```


```{R plot slopes}
### plot these slopes
ad.colors <- brewer.pal(9, "Paired")[c(4,7,3,6)]  ### numbers are one higher than admixture group because groups start at 0!
ad.names <- c("N.Swedish", "S.Swedish", "German", "Other")
p.pch <-c(17, 19)

plot.slope <- function(up.s, up.p, up.t){
  up.s <- up.s[4:1,]
  up.p <- up.p[4:1,]
  par(mar=c(4,4,2,2))
  plot(0,0, xlim=c(min(up.s), max(up.s)), ylim=c(0.5,nrow(up.s)), yaxt="n", ylab="", xlab="slope fitness~admix proportion", main=up.t)
  abline(v=0, col="grey70")
  for(up.r in c(1:4)){
    up.dat <- up.s[up.r,]
    up.pv <- up.p[up.r,]
    up.pv <- sapply(up.pv, function(x){x<0.05})
    up.pv <- as.numeric(up.pv)+1
    up.pv <- p.pch[up.pv]
    points(x=up.dat, y=rep(up.r,length(up.dat)), col=ad.colors, pch=up.pv, cex=1.7)
  }
  axis(2, at=1:4, labels=substr(rownames(up.s),1,3), las=2)
  legend(x="bottom",legend=c(ad.names, "sig", "not sig"), col=c(ad.colors, "black", "black"), pch=c(rep(19,4),19,17), ncol=3, bg="white")
}

par(mfcol=c(1,2))
plot.slope(up.s=slope1, up.p=pval1, up.t="2011")
plot.slope(up.s=slope2, up.p=pval2, up.t="2012")

pdf("./figures/slope.fitness.admixture.pdf", width=10, height=6)
par(mfcol=c(1,2))
plot.slope(up.s=slope1, up.p=pval1, up.t="2011")
plot.slope(up.s=slope2, up.p=pval2, up.t="2012")
dev.off()
```


```{R fitness rank}

#Is a pattern in proportion when accessions are ranked by beta?

pdf(file="./figures/fitness.rank.by.admix.prop.pdf", width=18, height=8)
par(mfcol=c(4,3), mar=c(0.2,4,1.5,0.2))
for(up.c in 16:23){
  p.dat <- a.fit[order(a.fit[,up.c]),]
  barplot(t(as.matrix(p.dat[,7:15])), col=brewer.pal(9, "Paired"), xaxt="n",space=0, ylab="genome proportion")
  mtext(colnames(a.fit)[up.c],side=3, line=0)
}
plot(0,type='n',axes=FALSE,ann=FALSE)
legend(legend=group.names, fill=brewer.pal(9, "Paired"), title="admixture group",bg="white",x="left")
#barplot(0, plot=FALSE, axes=FALSE, ann=FALSE,col=brewer.pal(9, "Paired"), legend.text=group.names, xaxt="n",space=0, args.legend=list(title="admixture group",bg="white",x="left") )

dev.off()
```

#### Test for simple effect of admixture group on fitness, using lines that are "well-assigned" to admixture groups (i.e. not considered admixed in 1001g data)

```{R admixture group fitness diffs}
ga.fit <- a.fit[a.fit$max.a>=0.60,]
ga.fit <- ga.fit[ga.fit$a.group%in%c(2,3,6),]
table(ga.fit$a.group)


#### 2011 first
ag1.aov <- as.list(1:4)
ag1.tukey <- as.list(1:4
                     )
col1 <- colnames(ga.fit)[grep("2011",colnames(ga.fit))]

for(up in 1:4){
  aov.up <- aov(get(col1[up])~as.factor(a.group), data=ga.fit)
  tt.up <- TukeyHSD(aov.up)
  ag1.aov[[up]] <- aov.up
  ag1.tukey[[up]] <- tt.up
}

par(mfrow=c(2,2))
lapply(ag1.tukey, plot)

lapply(ag1.aov,summary)


#### 2012 second
ag2.aov <- as.list(1:4)
ag2.tukey <- as.list(1:4)
col2 <- colnames(ga.fit)[grep("2012",colnames(ga.fit))]

for(up in 1:4){
  aov.up <- aov(get(col2[up])~as.factor(a.group), data=ga.fit)
  tt.up <- TukeyHSD(aov.up)
  ag2.aov[[up]] <- aov.up
  ag2.tukey[[up]] <- tt.up
}

par(mfrow=c(2,2))
lapply(ag2.tukey, plot)

lapply(ag2.aov,summary)

```

```{R plot by means rxn norm}
##reshape data
ga.long <- ga.fit[,c(1,6,grep("2011", colnames(ga.fit)), grep("2012", colnames(ga.fit)))]
mdata <- melt(ga.long, id=c("id","a.group"))
var.s <- strsplit(as.character(mdata$variable), split="_")
var.s <- do.call(rbind, var.s)
colnames(var.s) <- c("site", "year")
ga.long <- cbind(mdata, var.s)
colnames(ga.long)[4] <- "fitness"
ga.long$site = factor(ga.long$site,levels(ga.long$site)[c(1,2,4,3)])
ga.long$a.group = factor(ga.long$a.group, levels=c(2,6,3))

ag.colors <-group.colors[c(3,7,4)]

ad.group.boxplot <- ggplot(ga.long, aes(x=a.group, y=fitness, fill=as.factor(a.group))) +
  geom_violin() +
  geom_boxplot(width=0.1, show.legend=FALSE)+
  xlab("admixture group")+ 
  facet_grid(year~site) +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  scale_fill_manual(name = "admixture group", labels = c("German", "S.Swedish", "N.Swedish"), values=ag.colors) +
  scale_x_discrete(labels=c("German", "S.Swedish", "N.Swedish"))

print(ad.group.boxplot)

pdf(file="./figures/fitness.by.admix.group.pdf", width=8, height=6)
print(ad.group.boxplot)
dev.off()

### do this with N1-S2 names
ga.long.short <- ga.long
ga.long.short$site <- gsub("ADA","N2", ga.long.short$site)
ga.long.short$site <- gsub("RAM","N1", ga.long.short$site)
ga.long.short$site <- gsub("RAT","S2", ga.long.short$site)
ga.long.short$site <- gsub("ULL","S1", ga.long.short$site)

ad.group.short.boxplot <- ggplot(ga.long.short, aes(x=a.group, y=fitness, fill=as.factor(a.group))) +
  geom_violin() +
  geom_boxplot(width=0.1, show.legend=FALSE)+
  xlab("admixture group")+ 
  facet_grid(year~site) +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  scale_fill_manual(name = "admixture group", labels = c("German", "S.Swedish", "N.Swedish"), values=ag.colors) +
  scale_x_discrete(labels=c("German", "S.Swedish", "N.Swedish"))

pdf(file="./figures/fitness.by.admix.group.short.names.pdf", width=10, height=6)
print(ad.group.short.boxplot)
dev.off()
```


1.  In North, German lines always worse than both S and N Swedish.
2.  In South, 3/4 German and N Swedes worse than S. Sweden.  Exception is Rat - check this for slugs (where are sluggy lines?).
3.  Ullstorp, similar both years.
4.  In North, Switching effects of S. Swedes 2011/2012. (but N always better than German)

How does this translate cumulatively for 2 years???

### 2 year cumulative fitness
```{R cumul fitness data prep}
### need other component fitness traits
ows.dat <- read.table("./data/ows.fitted.values.txt", stringsAsFactors=FALSE)
sss.dat <- read.table("./data/sss.fitted.values.txt", stringsAsFactors=FALSE)
fec.dat <- read.table("./data/fecund.fitted.values.txt", stringsAsFactors=FALSE)
if(sum(ows.dat$id==sss.dat$id)==nrow(ows.dat)){print("ows.sss.OK")}
if(sum(ows.dat$id==sss.dat$id)==nrow(fec.dat)){print("fec.sss.OK")}
### these are all OK

cfit.a <- sss.dat$ADA_2011 * fec.dat$ADA_2011 * sss.dat$ADA_2012 * fec.dat$ADA_2012
cfit.r <- sss.dat$RAM_2011 * fec.dat$RAM_2011 * sss.dat$RAM_2012 * fec.dat$RAM_2012
cfit.u <- sss.dat$ULL_2011 * fec.dat$ULL_2011 * sss.dat$ULL_2012 * fec.dat$ULL_2012
cfit.t <- sss.dat$RAT_2011 * fec.dat$RAT_2011 * sss.dat$RAT_2012 * fec.dat$RAT_2012
cfit.dat <- cbind(ows.dat$id, cfit.a, cfit.r, cfit.u, cfit.t)
colnames(cfit.dat) <- c("id", "ADA", "RAM", "ULL", "RAT")


gac.fit <- merge(ga.fit,cfit.dat)
gac.long <- gac.fit[,c(1,6,25:28)]
gac.long<- melt(gac.long, id=c("id","a.group"))
colnames(gac.long)[3:4] <- c("site", "fitness")
```

```{R test cumulative fitness vs admix group by site}
cg1.aov <- as.list(1:4)
cg1.tukey <- as.list(1:4)

sites <- colnames(gac.fit)[25:28]

for(up in 1:4){
  aov.up <- aov(get(sites[up])~as.factor(a.group), data=gac.fit)
  tt.up <- TukeyHSD(aov.up)
  cg1.aov[[up]] <- aov.up
  cg1.tukey[[up]] <- tt.up
}

par(mfrow=c(2,2))
lapply(cg1.tukey, plot)

lapply(cg1.aov,summary)

```

```{R plot cumul fitness by admix group}
ad.group.cumul.boxplot <- ggplot(gac.long, aes(x=as.factor(a.group), y=fitness, fill=as.factor(a.group))) +
  geom_violin() +
  geom_boxplot(width=0.1, show.legend=FALSE)+
  xlab("admixture group")+ 
  facet_grid(~site) +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  scale_fill_manual(name = "admixture group", labels = c("Germany", "N.Sweden", "S.Sweden"), values=ag.colors) +
  scale_x_discrete(labels=c("Germany", "N.Sweden", "S.Sweden"))


print(ad.group.cumul.boxplot)

pdf(file="./figures/fitness.cumulative.by.admix.group.pdf", width=8, height=3.5)
print(ad.group.cumul.boxplot)
dev.off()



```


How about for the equivalant of the sampling from SnR???

### 1.5 year cumulative fitness
```{R snr relevant fitness data prep}

cfits.a <- sss.dat$ADA_2011 * fec.dat$ADA_2011 * sss.dat$ADA_2012 
cfits.r <- sss.dat$RAM_2011 * fec.dat$RAM_2011 * sss.dat$RAM_2012 
cfits.u <- sss.dat$ULL_2011 * fec.dat$ULL_2011 * sss.dat$ULL_2012 
cfits.t <- sss.dat$RAT_2011 * fec.dat$RAT_2011 * sss.dat$RAT_2012 
cfits.dat <- cbind(ows.dat$id, cfits.a, cfits.r, cfits.u, cfits.t)
colnames(cfits.dat) <- c("id", "ADA", "RAM", "ULL", "RAT")


snr.fit <- merge(ga.fit,cfits.dat)
snr.long <- snr.fit[,c(1,6,25:28)]
snr.long<- melt(snr.long, id=c("id","a.group"))
colnames(snr.long)[3:4] <- c("site", "fitness")
```

```{R test snr relevant fitness snr vs admix group by site}
snr.aov <- as.list(1:4)
snr.tukey <- as.list(1:4)

sites <- colnames(snr.fit)[25:28]

for(up in 1:4){
  aov.up <- aov(get(sites[up])~as.factor(a.group), data=snr.fit)
  tt.up <- TukeyHSD(aov.up)
  snr.aov[[up]] <- aov.up
  snr.tukey[[up]] <- tt.up
}

par(mfrow=c(2,2))
lapply(snr.tukey, plot)

lapply(snr.aov,summary)

```

```{R plot snr relevant fitness by admix group}
ad.group.cumul.boxplot <- ggplot(snr.long, aes(x=as.factor(a.group), y=fitness, fill=as.factor(a.group))) +
  geom_violin() +
  geom_boxplot(width=0.1)+
  xlab("admixture group")+ 
  facet_grid(~site) +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  scale_fill_manual(name = "admixture group", labels = c("Germany", "N.Sweden", "S.Sweden"), values=ag.colors) +
  scale_x_discrete(labels=c("Germany", "N.Sweden", "S.Sweden"))


print(ad.group.cumul.boxplot)

pdf(file="./figures/fitness.snr.relevant.by.admix.group.pdf", width=8, height=3.5)
print(ad.group.cumul.boxplot)
dev.off()



```

Check whether this fitness correlates with admixture group proportions, as well

```{R comp fit explain admix prop}

prop.lm <- function(fit.col, dat.frame){
  out.dat <- as.list(1:4)
  out.dat[[1]] <- lm(get(fit.col)~K9.ChrAll.3, data=dat.frame)
  out.dat[[2]] <- lm(get(fit.col)~K9.ChrAll.6, data=dat.frame)
  out.dat[[3]] <- lm(get(fit.col)~K9.ChrAll.2, data=dat.frame)
  p.nsw <- 1-(dat.frame$K9.ChrAll.3+dat.frame$K9.ChrAll.6+dat.frame$K9.ChrAll.2)
  out.dat[[4]] <- lm(get(fit.col)~p.nsw, data=dat.frame)
  return(out.dat)
}

cf.col <- colnames(gac.fit)[25:28]

#run models
adac <- prop.lm(fit.col=cf.col[1],dat.frame=gac.fit)
ramc <- prop.lm(fit.col=cf.col[2],dat.frame=gac.fit)
ullc <- prop.lm(fit.col=cf.col[3],dat.frame=gac.fit)
ratc <- prop.lm(fit.col=cf.col[4],dat.frame=gac.fit)

#get slopes
adac.s <- unlist(lapply(adac, function(x){summary(x)$coefficients[2,1]}))
ramc.s <- unlist(lapply(ramc, function(x){summary(x)$coefficients[2,1]}))
ullc.s <- unlist(lapply(ullc, function(x){summary(x)$coefficients[2,1]}))
ratc.s <- unlist(lapply(ratc, function(x){summary(x)$coefficients[2,1]}))
slopec <- rbind(adac.s, ramc.s, ullc.s, ratc.s)
colnames(slopec) <- c("n.swedish", "s.swedish", "german", "other")


#get pvals
adac.p <- unlist(lapply(adac, function(x){summary(x)$coefficients[2,4]}))
ramc.p <- unlist(lapply(ramc, function(x){summary(x)$coefficients[2,4]}))
ullc.p <- unlist(lapply(ullc, function(x){summary(x)$coefficients[2,4]}))
ratc.p <- unlist(lapply(ratc, function(x){summary(x)$coefficients[2,4]}))
pvalc <- rbind(adac.p, ramc.p, ullc.p, ratc.p)
colnames(pvalc) <- c("n.swedish", "s.swedish", "german", "other")


```


```{R plot composite slopes}
### plot these slopes
ad.colors <- brewer.pal(9, "Paired")[c(4,7,3,6)]  ### numbers are one higher than admixture group because groups start at 0!
ad.names <- c("N.Swedish", "S.Swedish", "German", "Other")
p.pch <-c(17, 19)

plot.slope <- function(up.s, up.p, up.t){
  up.s <- up.s[4:1,]
  up.p <- up.p[4:1,]
  par(mar=c(4,4,2,2))
  plot(0,0, xlim=c(min(up.s), max(up.s)), ylim=c(0.5,nrow(up.s)), yaxt="n", ylab="", xlab="slope fitness~admix proportion", main=up.t)
  abline(v=0, col="grey70")
  for(up.r in c(1:4)){
    up.dat <- up.s[up.r,]
    up.pv <- up.p[up.r,]
    up.pv <- sapply(up.pv, function(x){x<0.05})
    up.pv <- as.numeric(up.pv)+1
    up.pv <- p.pch[up.pv]
    points(x=up.dat, y=rep(up.r,length(up.dat)), col=ad.colors, pch=up.pv, cex=1.7)
  }
  axis(2, at=1:4, labels=substr(rownames(up.s),1,3), las=2)
  legend(x="bottom",legend=c(ad.names, "sig", "not sig"), col=c(ad.colors, "black", "black"), pch=c(rep(19,4),19,17), ncol=3, bg="white")
}

plot.slope(up.s=slopec, up.p=pvalc, up.t="Composite")

pdf("./figures/slope.composite.fitness.admixture.pdf", width=10, height=6)
plot.slope(up.s=slopec, up.p=pvalc, up.t="Composite")
dev.off()
```

Let's try the relationship between fitness and genome proportion taking out non-admixed individuals

```{R lm fit admix by prop}
admix.fit <- a.fit[a.fit$max.a<=0.95,]
fit.data <- admix.fit


prop.lm.a <- function(fit.col, fit.data){
  out.dat <- as.list(1:3)
  out.dat[[1]] <- lm(get(fit.col)~K9.ChrAll.3, data=fit.data)
  out.dat[[2]] <- lm(get(fit.col)~K9.ChrAll.6, data=fit.data)
  out.dat[[3]] <- lm(get(fit.col)~K9.ChrAll.2, data=fit.data)
  return(out.dat)
}


### 2011 first
fit.columns <- colnames(admix.fit)[grep("2011",colnames(a.fit))]
ada1a <- prop.lm.a(fit.col=fit.columns[1], fit.data=admix.fit)
ram1a <- prop.lm.a(fit.col=fit.columns[2], fit.data=admix.fit)
ull1a <- prop.lm.a(fit.col=fit.columns[3], fit.data=admix.fit)
rat1a <- prop.lm.a(fit.col=fit.columns[4], fit.data=admix.fit)

### 2012 second
fit.columns <- colnames(admix.fit)[grep("2012",colnames(a.fit))]
ada2a <- prop.lm.a(fit.col=fit.columns[1], fit.data=admix.fit)
ram2a <- prop.lm.a(fit.col=fit.columns[2], fit.data=admix.fit)
ull2a <- prop.lm.a(fit.col=fit.columns[3], fit.data=admix.fit)
rat2a <- prop.lm.a(fit.col=fit.columns[4], fit.data=admix.fit)

###### slopes ######
### 2011
ada1.sa <- unlist(lapply(ada1a, function(x){summary(x)$coefficients[2,1]}))
ram1.sa <- unlist(lapply(ram1a, function(x){summary(x)$coefficients[2,1]}))
ull1.sa <- unlist(lapply(ull1a, function(x){summary(x)$coefficients[2,1]}))
rat1.sa <- unlist(lapply(rat1a, function(x){summary(x)$coefficients[2,1]}))
slope1a <- rbind(ada1.sa, ram1.sa, ull1.sa, rat1.sa)
colnames(slope1a) <- c("N.Swedish", "S.Swedish", "German")

### 2012
ada2.sa <- unlist(lapply(ada2a, function(x){summary(x)$coefficients[2,1]}))
ram2.sa <- unlist(lapply(ram2a, function(x){summary(x)$coefficients[2,1]}))
ull2.sa <- unlist(lapply(ull2a, function(x){summary(x)$coefficients[2,1]}))
rat2.sa <- unlist(lapply(rat2a, function(x){summary(x)$coefficients[2,1]}))
slope2a <- rbind(ada2.sa, ram2.sa, ull2.sa, rat2.sa)
colnames(slope2a) <- c("N.Swedish", "S.Swedish", "German")

######### pvals #############
### 2011
ada1.pa <- unlist(lapply(ada1a, function(x){summary(x)$coefficients[2,4]}))
ram1.pa <- unlist(lapply(ram1a, function(x){summary(x)$coefficients[2,4]}))
ull1.pa <- unlist(lapply(ull1a, function(x){summary(x)$coefficients[2,4]}))
rat1.pa <- unlist(lapply(rat1a, function(x){summary(x)$coefficients[2,4]}))
pval1a <- rbind(ada1.pa, ram1.pa, ull1.pa, rat1.pa)
colnames(pval1a) <- c("N.Swedish", "S.Swedish", "German")

### 2012
ada2.pa <- unlist(lapply(ada2a, function(x){summary(x)$coefficients[2,4]}))
ram2.pa <- unlist(lapply(ram2a, function(x){summary(x)$coefficients[2,4]}))
ull2.pa <- unlist(lapply(ull2a, function(x){summary(x)$coefficients[2,4]}))
rat2.pa <- unlist(lapply(rat2a, function(x){summary(x)$coefficients[2,4]}))
pval2a <- rbind(ada2.pa, ram2.pa, ull2.pa, rat2.pa)
colnames(pval2a) <- c("N.Swedish", "S.Swedish", "German")
```


```{R admix indiv. admix by beta plot}

up.site="ULL_2011"
ad.groups <- c("K9.ChrAll.2","K9.ChrAll.6","K9.ChrAll.3")


out.dat <- as.list(1:3)
for(up in 1:3){
  up.group <- ad.groups[up]
  up.dat <- admix.fit[,c(grep(up.site, colnames(admix.fit)),which(colnames(admix.fit)==up.group))]
  up.dat$ad.group <- up.group
  colnames(up.dat) <- c("fitness","ad.prop","ad.group")
  out.dat[[up]] <- up.dat
}
out.dat <- do.call(rbind, out.dat)
out.dat$ad.group <- factor(out.dat$ad.group, levels=ad.groups)
out.dat <- out.dat[out.dat$ad.prop>0.05,]

ggplot(out.dat, aes(ad.prop, fitness, colour=ad.group)) +
  geom_point()+
  geom_smooth(method="lm") +
  scale_color_manual(values=ag.colors)

```

Hmmmm...not sure about this.  In any case, it needs a better model (proportions are not independent - what happens if you have 50% s + 50% n versus 50s + 50g?)
